backend/requirements.txt
# Web framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0

# Database
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
pgvector>=0.3,<1.0
alembic>=1.14,<2.0

# Task queue
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0

# Config
pydantic-settings>=2.7,<3.0

# Auth / Security
bcrypt>=4.0,<5.0
python-jose[cryptography]>=3.3,<4.0
slowapi>=0.1.9,<1.0
email-validator>=2.0,<3.0

# Scheduler
apscheduler>=3.10,<4.0

# AI / Matching
numpy>=1.26,<3.0
langdetect>=1.0.9,<2.0

# HTTP client
httpx>=0.28,<1.0

# Dev / Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.25,<1.0
httpx  # also used for TestClient
ruff>=0.9,<1.0
